### 2021-02-26 #3069 @whedon
[PRE REVIEW]: coxeter: A Python package for working with Shapes
**Submitting author:** @vyasr (<a href="http://orcid.org/0000-0001-5181-9532">Vyas Ramasubramani</a>)
**Repository:** <a href="https://github.com/glotzerlab/coxeter" target ="_blank">https://github.com/glotzerlab/coxeter</a>
**Version:** 0.5.0
