### 2021-03-08 #3069 @jni
@kthyng I just finished my last review, and this is in my wheelhouse, so I'd be happy to take this on.

@vyasr if you have suggestions for reviewers that don't [have a conflict of interest per the JOSS guidelines](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy), please let us know!
