### 2021-03-09 #3069 @Jacob-Barhak
Hi Vyas,

This paper looks interesting - if you want my review - I will make it
publicly below

I read through, I saw many claims that the thing I missed most was an
example. If the Authors would give a simple example and show how they apply
their claims to it it will give a good idea to what they are doing.

Calculating volumes for examples has been practiced in CAD systems for many
years - algo using all sorts of constructions from primities with
techniques like CSG trees. I would not comprehend from the text why this is
such a difficult problem - perhaps if the authors show an example of why it
is difficult and how they do this efficiently, it would help. I could not
understand if shape generation is similar to CSG trees or something else.

I did not look at the code, not checked all the references.

I am curious to know how this can help biological modeling - if the simple
example can include this, it would be great.

Generally I like the idea of more python libraries for geometry - the more
the better.  I liked the idea of having a library of primitive shapes with
properties - if it is open to contributions, even better, in which case,
please provide information on how people can contribute.

More information on the paper will be great. Yet I also liked that it was
short and a quick read. So The authors will have to find a balance. Perhaps
links to examples may be a good solution here.

Jacob Barhak Ph.D.
https://sites.google.com/view/jacob-barhak/home

On Mon, Mar 8, 2021 at 5:25 PM Vyas Ramasubramani <notifications@github.com>
wrote:

> @jni <https://github.com/jni> thanks for touching base. Looking through
> the spreadsheet of reviewers @amritagos <https://github.com/amritagos>,
> @Jacob-Barhak <https://github.com/Jacob-Barhak>, and @tmaric
> <https://github.com/tmaric> are three people who might be interested.
>
