### 2021-03-09 #3069 @jni
@Jacob-Barhak thank you for jumping right in, but review hasn't started! :sweat_smile: I should have been more clear above. Review proceeds in a new issue where you would get a reviewer checklist, along with clear criteria for review. If you agree to be a reviewer, the guide is here

https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
https://joss.readthedocs.io/en/latest/review_criteria.html

The review checklist, which again will appear in the issue as actual tick-boxes, is here:

https://joss.readthedocs.io/en/latest/review_checklist.html

JOSS reviews are focused on ensuring software functionality, proper documentation, and proper guidelines for community participation (bug reports, code contributions) in the software.

Given the above, please let me know if you would agree to review the paper; that would be wonderful!

@amitragos, @tmaric, would either of you be willing to help review this paper?

Thank you!
