### 2021-03-09 #3069 @Jacob-Barhak
So Juan,

You will find my opinion in the paper posted already. I will go no further
than this, especially after seeing the links you sent me.

You see, there is a difference between giving an opinion and performing
tests on a system to give it approval by guidelines one does not agree
with. You went a bit too far with those guidelines.

Especially I did not like the channeling of the license to OSI approved
licenses. You see OSI should no longer dictate licenses and I cannot find
myself in true conscience following their guidelines. I was once a member
of OSI, yet no longer can be affiliated. A few years ago they rejected CC0
as a license and I learned about it this year. This license is more open
than many of the copyright based licenses that OSI approves. In fact
the term open source is misleading and more restrictive if you follow some
of the licenses that OSI approves. I tried to bring it to debate within an
OSI conference without success.

You will find a discussion on the topic here:

Jacob. Barhak, Open Source and Sustainability, COMBINE 2020 October 5-9. Video:
https://drive.google.com/drive/folders/1actGnx6FwvoCcPrrF3qbnO0AmHt10WN6
starting from minute 13:10. Presentation:
*https://jacob-barhak.github.io/COMBINE2020_OpenSource_upload_2020_10_04.odp
<https://jacob-barhak.github.io/COMBINE2020_OpenSource_upload_2020_10_04.odp>*

This alone means I cannot abide by your guidelines that force an OSI
approved license.

However, I also looked at your Conflict of Interest statement - it seems
someone needs to look at the definition of conflict of interest. Conflict
of Interest is resolved by public declaration of the conflict - if the
review is public and non blind, this alone kind of removes the conflict of
interest since people can check. Therefore I am unclear why you
redefined it at all.

In fact, following OSI guidelines will put me in conflict of interest by
its true definition since I personally have potential monetary gains from
patents.

I also looked at your board members and found out that it includes at least
one person who did not support a public open review policy in the past
which I promote.

Finally, the BDFL title seems outdated - I actually raised an issue about
that.

Considering all those considerations I understand that your Journal is
going in the wrong direction. Your people may mean well, yet it seems they
do not comprehend the path this leads to. It is time to challenge this, and
this is what I am doing here.

Hopefully you will understand my hints and improve things -- I will also
raise the issue to NumFocus.

It is important those issues are discussed openly and publicly - I believe
things can still be remedied.

Jacob Barhak
Sole Proprietor

On Tue, Mar 9, 2021 at 6:05 AM Juan Nunez-Iglesias <notifications@github.com>
wrote:

> @Jacob-Barhak <https://github.com/Jacob-Barhak> thank you for jumping
> right in, but review hasn't started! 😅 I should have been more clear
> above. Review proceeds in a new issue where you would get a reviewer
> checklist, along with clear criteria for review. If you agree to be a
> reviewer, the guide is here
>
> https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
> https://joss.readthedocs.io/en/latest/review_criteria.html
>
> The review checklist, which again will appear in the issue as actual
> tick-boxes, is here:
>
> https://joss.readthedocs.io/en/latest/review_checklist.html
>
> JOSS reviews are focused on ensuring software functionality, proper
> documentation, and proper guidelines for community participation (bug
> reports, code contributions) in the software.
>
> Given the above, please let me know if you would agree to review the
> paper; that would be wonderful!
>
> @amritragos, @tmaric <https://github.com/tmaric>, would either of you be
> willing to help review this paper?
>
> Thank you!
>
