### 2021-03-10 #3069 @jni
@wolfv I wonder if as author of scikit-geometry you might be able/willing to review this submission? Please see [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html) for our reviewer guide and feel free to ask below if you have any questions about the process. Essentially, the review is focused on ensuring functionality, documentation, and community around the software described by the paper. Thanks!
