### 2021-03-10 #3069 @Jacob-Barhak
So Juan,

If this is not the proper venue please direct the discussion to the proper
venue. Hopefully a good public discussion can remedy things.

If you believe there is value in the venue, then the discussion
should continue somewhere to improve things - hopefully you have a
suggestion. If you cannot suggest one, since this venue triggered the issue
and it is public, it would be good enough. Yet I do hope you have a public
redirect to channel things properly.

Jacob

On Wed, Mar 10, 2021 at 4:40 PM Juan Nunez-Iglesias <
notifications@github.com> wrote:

> I'm sorry you feel that these policies disqualify us, @Jacob-Barhak
> <https://github.com/Jacob-Barhak>. Not all JOSS editors subscribe to all
> JOSS policies, but we believe the journal is on the whole a net good for
> the community. At any rate, I'd like to keep this issue focused on the
> pre-review of this paper. Thanks for your thoughts.
>
> @tmaric <https://github.com/tmaric> thank you so much for agreeing to
> review!
>
