### 2021-03-10 #3069 @jni
Yeah I actually looked around for an appropriate forum but couldn't exactly find one. Our pre-submission enquiries go to https://github.com/openjournals/joss, so creating a new issue there is probably the best bet. I would recommend separate issues for CoI and license requirements discussions.
