### 2021-03-11 #3069 @Jacob-Barhak
Thanks Juan,

The license issue is now listed here:
https://github.com/openjournals/joss/issues/889

The discussion should continue in that link.

The other issues are minor compared to this one. Hopefully it will gain
attention and will be resolved.

Thank you for redirecting

Jacob

On Wed, Mar 10, 2021 at 5:00 PM Juan Nunez-Iglesias <
notifications@github.com> wrote:

> Yeah I actually looked around for an appropriate forum but couldn't
> exactly find one. Our pre-submission enquiries go to
> https://github.com/openjournals/joss, so creating a new issue there is
> probably the best bet. I would recommend separate issues for CoI and
> license requirements discussions.
>
