### 2021-03-11 #3098 @whedon
[REVIEW]: coxeter: A Python package for working with Shapes
**Submitting author:** @vyasr (<a href="http://orcid.org/0000-0001-5181-9532">Vyas Ramasubramani</a>)
**Repository:** <a href="https://github.com/glotzerlab/coxeter" target ="_blank">https://github.com/glotzerlab/coxeter</a>
**Version:** v0.6.1

@tmaric, @wolfv please carry out your review in this issue by updating the checklist below. If you cannot edit the checklist please:
