### 2021-03-11 #3098 @jni
@tmaric @wolfv, thank you again for volunteering to review, much appreciated! :pray: In case this is your first time here, JOSS reviews have a nice, guided structure described in our [review guidelines](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html), [review criteria](https://joss.readthedocs.io/en/latest/review_criteria.html), and the checklist above. Reviews are collaborative, much like code review. If you have any questions about the process, don't hesitate to ask here.

Thanks again!
