### 2021-04-14 #3098 @wolfv
Hi, so finally found some time to read the paper and take a look at the docs + code.

I think this is excellent work!

I have some comments to the documentation of this project:

- it would be great to see some 'real-life' example use (maybe in a jupyter notebook, with visible plots of the different shapes, too)
- I think the docs are very thorough, but a tutorial or some other high-level introduction seems to be missing (might go hand-in-hand with the real-life example mentioned above).

The paper reads great and the authors demonstrate a clear need for a software like this.
Does this make sense as a review? It's my first time doing a JOSS review, so let me know if I need to add more review notes.
