### 2021-04-14 #3098 @jni
Yes, good tips @wolfv and thank you for taking the time to look over everything!

The only question is whether you would suggest this as feedback for the authors but still recommend acceptance, or whether you see a tutorial/real-life example as a pre-requisite before publication. I can see arguments either way.
