### 2021-04-15 #3098 @vyasr
Yes thanks for the review @wolfv. Whether or not you think this is a prerequisite for publication, a tutorial+examples is definitely something I'll work to add. What would you think about publishing the notebooks used to generate the components of the paper figure as a set of representative "real-life" examples? Each panel in the figure was made with a separate notebook that I could directly add (probably most appropriate after the paper is published and I can post slightly modified versions to avoid any copyright issues).
