### 2021-04-15 #3098 @jni
@vyasr since you are the author you can choose whatever copyright you want. If you use CC-BY/CC0 for your docs and refer to them in the paper, everything should be fine. We don't publish papers that have been published in other peer-reviewed venues, but basing the paper on the docs and vice versa is A-ok.
