### 2021-04-25 #3098 @vyasr
Conversely, since the paper will be published under CC-BY 4.0 (at least the last one that I published here was), it should also be fine to go in the opposite direction and cite the paper in the docs after it's been published, correct? Assuming @wolfv is fine to publish the paper without these examples already in place, I'll probably wait until this review is done then do a an update to the docs both to add references to this paper and add the new examples.
