### 2021-05-16 7ba348a
Use nbsphinx to include new notebook in the documentation.
