+++ doc/source/examples.rst
@@ -1,0 +1,8 @@
+Examples
+========
+
+While **coxeter**'s API is very easy to work with, it can be helpful to see some real demonstrations of how it can be used.
+Here we include some practical examples of using **coxeter**.
+
+.. toctree::
+   examples/PointInShape.ipynb
