+++ doc/source/examples.rst
@@ -7,3 +7,4 @@ Here we include some practical examples of using **coxeter**.
 .. toctree::
    examples/PointInShape.ipynb
    examples/DistanceToSurface.ipynb
+   examples/InertiaTensors.ipynb
