### 2021-05-20 bfd0e2b
Add insphere/circumsphere example.
