+++ doc/source/examples.rst
@@ -9,3 +9,4 @@ Here we include some practical examples of using **coxeter**.
    examples/DistanceToSurface.ipynb
    examples/InertiaTensors.ipynb
    examples/Spheres.ipynb
+   examples/FormFactors.ipynb
