### 2021-05-20 4b10dd4
Add nbqa to pre-commit and validate new doc examples (and paper).
