+++ .pre-commit-config.yaml
@@ -32,3 +32,16 @@ repos:
         entry: flake8
         language: python
         types: [python]
+  - repo: https://github.com/nbQA-dev/nbQA
+    rev: 0.8.1
+    hooks:
+      - id: nbqa-pyupgrade
+        args:
+          - --nbqa-mutate
+          - --py36-plus
+      - id: nbqa-isort
+        args:
+          - --nbqa-mutate
+      - id: nbqa-black
+        args:
+          - --nbqa-mutate
