### 2021-05-20 #3098 @vyasr
I've opened a PR adding simplified examples based on the figures as per @wolfv's suggestion!
