### 2021-05-20 e86d25b
Update requirements for docs.
