### 2021-05-21 efb4c4e
Try installing jupyter.
