+++ doc/requirements.txt
@@ -5,3 +5,4 @@ sphinxcontrib-bibtex>=2.0.0
 autodocsumm
 nbsphinx
 fresnel
+jupyter
