+++ readthedocs.yml
@@ -2,3 +2,5 @@ python:
    version: 3
    install:
        - requirements: docs/requirements.txt
+       - method: pip
+         path: .
