+++ readthedocs.yml
@@ -1,6 +1,8 @@
+version: 2
+
 python:
    version: 3
    install:
-       - requirements: docs/requirements.txt
        - method: pip
          path: .
+       - requirements: docs/requirements.txt
