+++ readthedocs.yml
@@ -5,4 +5,4 @@ python:
    install:
        - method: pip
          path: .
-       - requirements: docs/requirements.txt
+       - requirements: doc/requirements.txt
