### 2021-05-21 5aa4216
Switch to conda RTD to get fresnel.
