### 2021-06-14 d9c74f7
Add message about supported backends/primitives.
