### 2021-06-14 54a9ffd
Use scene_kwargs instead of **scene_kwargs.
