### 2021-06-14 a62d01c
Refactor backend/scene.
