### 2021-06-15 43bc1f2
Clarify dimensionality.
