### 2021-06-24 #3098 @jni
@evetion thank you so much for agreeing to review this submission! 🙏 You should have got an invitation to join the repo, which you can check at https://github.com/openjournals/joss-reviews/invitations. Accepting the invitation will allow you to tick off boxes in the checklist at the top of this issue. Please let me know if you don't have access and I'll reinvite you.

JOSS reviews happen in public. You can add comments to this issue to raise questions with the authors, or create issues on the original software repo — whichever you find more appropriate for a particular discussion.

Please let me know if you have any more questions about the process!
