### 2021-06-24 1ebd799
Apply suggestions from code review
