### 2021-06-30 #3098 @vyasr
@evetion thanks for the review! Apologies for the slow response, I'm in the middle of a few things but will get back to you ASAP. It looks like one of my coauthors already helped to start addressing your issues.

Regarding how I found you, you're on [the list of potential JOSS reviewers](https://docs.google.com/spreadsheets/d/1PAPRJ63yq9aPC1COLjaQp8mHmEq3rZUzwUYxTulyu78/edit#gid=856801822). I know that I added myself at one point at someone's request, possibly around when I first submitted a paper to JOSS.
