### 2021-07-08 e6043ce
Retitle building docs section of README.
