### 2021-07-08 #3098 @evetion
@vyasr The recent activity on the issues over at coxeter is great, but note that most of those are nice to haves and not required for joss. In case this wasn't clear, for this review to succeed, there are two open *blocking questions* in [my comment above](https://github.com/openjournals/joss-reviews/issues/3098#issuecomment-868688025), namely the statement of need in the documentation and the question of authorship.
