### 2021-07-08 #3098 @vyasr
@evetion thanks for the ping on that. I'd appreciate any feedback from you on my work on the non-blocking items as well, but regarding the blockers:
- As I'm sure you noticed, Sharon Glotzer has not contributed any code to coxeter, but she has been involved in all of the research that led to its creation and helped guide the creation of prototypes for many features that eventually made their way into the package. Some concrete examples of those features that eventually found their way into the package:
    - Calculating diffraction patterns on highly anisotropic building blocks was a major focus of a dissertation published in our group a number of years ago. Sharon was involved in determining the necessity and the means for incorporating anisotropic form factors into this calculation, which led to the initial prototypes for calculating form factors that I worked into for coxeter.
    - Some of the research that Sharon and I worked on involves studying the dynamics of anisotropic particles. Those types of simulations require precise calculations of the inertia tensor of a shape in order to reproduce accurate dynamics, and we found that inaccurate calculations were leading to incorrect results. That drove the implementation of the more robust triangulation and inertia tensor calculations in coxeter.
    - Another research area that Sharon has recently been pushing is in theoretical approaches to anisotropic particles based on perturbation theories. Many of those perturbation theories rely on treating shapes relative to their inspheres/incircles, which means we needed robust ways to estimate the insphere radii of shapes as well as the relative dilation at a particular point on the surface. That was another calculation that ended up in coxeter.
- Some of my recent activity was targeted at your request for a more visible statement of need, specifically [the changes to the README](https://github.com/glotzerlab/coxeter/pull/155/files#diff-7b3ed02bc73dc06b7db906cf97aa91dec2b2eb21f2d92bc5caa761df5bbc168f). Do you think this is enough information, or would you prefer that I inserted more of the statement of need from the paper draft directly into the paper?
