### 2021-07-11 #3098 @evetion
> @evetion fantastic! Thank you for that review! Can you confirm that you are ready for the paper to be accepted then?

Yep, I confirm we can accept this paper. 🎉
