### 2021-07-12 96eb5ab
Apply suggestions from code review
Co-authored-by: Bradley Dice <bdice@bradleydice.com>
