+++ doc/source/quickstart.rst
@@ -3,14 +3,14 @@
 Quickstart Tutorial
 ===================
 
-Once you have :ref:`installed <installing>` **coxeter**, most workflows involve creating an instance of a shape:
+Once you have :ref:`installed <installing>` **coxeter**, most workflows involve creating an instance of a shape, such as a :class:`~coxeter.shapes.Polygon`:
 
 .. code-block:: python
 
     >>> import coxeter
     >>> square = coxeter.shapes.Polygon([[0, 0], [1, 0], [1, 1], [0, 1]])
 
-All shapes may be found in the ``coxeter.shapes`` subpackage and are created in a similar manner.
+All shapes may be found in the `coxeter.shapes` subpackage and are created in a similar manner.
 For instance, making a sphere requires a radius: ``sphere = coxeter.shapes.Sphere(3)``.
 Once you have a shape, you can immediately query it for properties.
 
@@ -26,7 +26,8 @@ Once you have a shape, you can immediately query it for properties.
     >>> square.perimeter
     4.0
 
-**coxeter** comes with a large range of *shape families*, collections of shapes with standard definitions so that you don't have to parameterize them yourself:
+The **coxeter** library comes with a range of *shape families*, collections of shapes with standard definitions so that you don't have to parameterize them yourself.
+The regular :math:`n`-gon family is such an example, provided as :class:`coxeter.families.RegularNGonFamily`.
 
 .. code-block:: python
 
@@ -40,8 +41,8 @@ Once you have a shape, you can immediately query it for properties.
            [ 0.31, -0.54,  0.  ]])
 
 Part of what makes **coxeter** so powerful is that all shapes are mutable.
-This means that once you have a prototype of a shape it can be modified to fit a specific need.
-For example, the snippet below finds the area of the smallest regular pentagon that contains a unit area equilateral triangle:
+This means that once you have a prototype of a shape, it can be modified to fit a specific need.
+For example, the snippet below finds the area of the smallest regular pentagon that contains an equilateral triangle of unit area:
 
 .. code-block:: python
 
