### 2021-07-13 #3098 @jni
@vyasr Things are looking good! Could you please:

- make a new tagged release on github
- make a new Zenodo archive (ensure that the title matches the paper title)
- report the DOI and version number here

Thank you!
