### 2021-07-14 5a7b08e
Update changelog and fix minor errors.
