### 2021-07-14 bc13077
Make middle names consistent in paper.
