### 2021-07-14 5d4c650
Bump up to version 0.6.0.
