+++ coxeter/__init__.py
@@ -21,4 +21,4 @@ from .shape_getters import from_gsd_type_shapes
 
 __all__ = ["families", "shapes", "from_gsd_type_shapes"]
 
-__version__ = "0.5.0"
+__version__ = "0.6.0"
