### 2021-07-14 #3098 @vyasr
@jni Here's the [new tagged release on Github](https://github.com/glotzerlab/coxeter/releases/tag/v0.6.0) and the [link for PyPI](https://pypi.org/project/coxeter/). The Zenodo archive is [here](https://zenodo.org/record/5102768#.YO8CuW5lCCM) for the current version, which is now 0.6.0, and the DOI is 10.5281/zenodo.5102768. Thanks!
