### 2021-07-14 e047ec6
Remove space in bib file.
