+++ paper/paper.bib
@@ -198,7 +198,7 @@
   publisher = {American Physical Society},
   doi = {10.1103/PhysRevLett.107.135701},
 }
-@article {Rossi2015,
+@article{Rossi2015,
 	author = {Rossi, Laura and Soni, Vishal and Ashton, Douglas J. and Pine, David J. and Philipse, Albert P. and Chaikin, Paul M. and Dijkstra, Marjolein and Sacanna, Stefano and Irvine, William T. M.},
 	title = {Shape-sensitive crystallization in colloidal superball fluids},
 	volume = {112},
