### 2021-07-14 #3098 @vyasr
@jni quick heads up, my coauthor @bdice noticed a very minor typo in one of the citations that I just fixed on our main branch. Those changes came after the release and the Zenodo archive, hope that's not an issue.
