### 2021-07-15 #3098 @jni
@vyasr it's not an issue in that it's easily fixable, but unfortunately we will need you to tag/release/archive again, sorry about that! 0.6.1? :grimacing:
