+++ setup.py
@@ -12,7 +12,7 @@ try:
 except ImportError:
     readme = desc
 
-version = "0.6.0"
+version = "0.6.1"
 
 
 ################################################
